Fixed 2 errors on generation of SQL expression for Sphinx.
There was an error generating the SQL expression for Sphinx
caused by assuming that query.query generated valid SQL.
This commit uses mogrify from psycopg for Postgres
and the same code pymysql uses for Mysql to generate the
correct SQL.

There was an error where the indexes were not configured when not
imported in the main code. This commit uses Django's new API
for Apps to import the indexes in sphinxql.ready().

In Mysql, generating SQL expression now requires DB access since
the escaping algorithm depends on the DB settings.

This commit also simplifies the configuration code.